pyproject.toml
[tool.poetry]
name = "evacode-backend"
version = "0.1.0"
description = ""
authors = ["alex <sashason1990@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
django-taggit = "^5.0.1"
django-ckeditor = "^6.7.1"
pillow = "^10.2.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
